% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculos.r
\name{calcular_overlap_tippett}
\alias{calcular_overlap_tippett}
\title{Calcular área de solapamiento para gráficas Tippett}
\usage{
calcular_overlap_tippett(base, var_llr, var_categoria)
}
\arguments{
\item{base}{Data frame con los datos}

\item{var_llr}{Nombre de la columna con valores de LLR}

\item{var_categoria}{Nombre de la columna con categorías}
}
\value{
Área de solapamiento
}
\description{
Calcula el área de solapamiento entre las curvas acumuladas
de las hipótesis en una gráfica Tippett.
}
\examples{
\dontrun{
overlap_tip <- calcular_overlap_tippett(mis_datos, "llr_valores", "categoria")
}
}
