% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecciones.r
\name{encontrar_cruce}
\alias{encontrar_cruce}
\title{Encontrar punto de cruce en gráfica Tippett}
\usage{
encontrar_cruce(base, var_llr, var_categoria)
}
\arguments{
\item{base}{Data frame con los datos}

\item{var_llr}{Nombre de la columna con valores de LLR}

\item{var_categoria}{Nombre de la columna con categorías}
}
\value{
Valor numérico del punto de cruce en escala log10
}
\description{
Encuentra el valor de x donde se intersectan las curvas SS y DS
en una gráfica Tippett.
}
\examples{
\dontrun{
cruce <- encontrar_cruce(datos, "llr_valores", "categoria")
}
}
