% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extraer_nombre_completo}
\alias{extraer_nombre_completo}
\title{Extraer nombre completo de archivos de audio}
\usage{
extraer_nombre_completo(texto)
}
\arguments{
\item{texto}{Vector de caracteres con nombres de archivo}
}
\value{
Vector de caracteres con nombres limpios
}
\description{
Extrae el nombre base eliminando sufijos comunes en archivos de audio forense.
}
\examples{
extraer_nombre_completo("audio_HESP_T2.wav")
}
