% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculos.r
\name{fusion_LR}
\alias{fusion_LR}
\title{Fusionar múltiples LRs}
\usage{
fusion_LR(data, target_var, predictor_vars)
}
\arguments{
\item{data}{Data frame con los datos}

\item{target_var}{Variable objetivo binaria (1 = misma persona, 0 = diferente)}

\item{predictor_vars}{Vector con nombres de las variables predictoras}
}
\value{
Vector con valores fusionados (log10 del LR calibrado)
}
\description{
Combina múltiples variables de LLR mediante regresión logística
para producir un valor fusionado.
}
\examples{
\dontrun{
fusionados <- fusion_LR(datos, "misma_persona_binaria", c("llr_10", "llr_20", "llr_30"))
}
}
