% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar.r
\name{generar_excel_resultados}
\alias{generar_excel_resultados}
\title{Generar Excel con resultados completos}
\usage{
generar_excel_resultados(
  base_datos,
  columna_llr = "llr_40",
  columna_grupo = "misma_persona",
  valores_n = seq(5, 35, by = 5),
  nombre_archivo = "Resultados-completos.xlsx",
  directorio = resultados_analisis
)
}
\arguments{
\item{base_datos}{Data frame con los datos}

\item{columna_llr}{Nombre de la columna con valores de LLR}

\item{columna_grupo}{Nombre de la columna con grupos}

\item{valores_n}{Vector con valores de n a analizar}

\item{nombre_archivo}{Nombre del archivo Excel a generar}

\item{directorio}{Directorio donde guardar el archivo}
}
\value{
Lista invisible con los data frames generados
}
\description{
Genera un archivo Excel con dos hojas: intersecciones y Cllr
para diferentes valores de n.
}
\examples{
\dontrun{
generar_excel_resultados(mis_datos, "llr_40", "misma_persona",
                         valores_n = seq(5, 35, by = 5))
}
}
