% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar.r
\name{generar_excel_resultados_15_min}
\alias{generar_excel_resultados_15_min}
\title{Generar Excel con resultados (versión 15 minutos)}
\usage{
generar_excel_resultados_15_min(
  base_datos,
  columna_llr = "llr_40",
  columna_grupo = "misma_persona",
  valores_n = seq(5, 35, by = 5),
  nombre_archivo = "Resultados-completos.xlsx",
  directorio = resultados_analisis
)
}
\arguments{
\item{base_datos}{Data frame con los datos}

\item{columna_llr}{Nombre de la columna con valores de LLR}

\item{columna_grupo}{Nombre de la columna con grupos}

\item{valores_n}{Vector con valores de n a analizar}

\item{nombre_archivo}{Nombre del archivo Excel a generar}

\item{directorio}{Directorio donde guardar el archivo}
}
\value{
Lista invisible con los data frames generados
}
\description{
Versión especializada que incluye categorías de duración promedio.
}
\examples{
\dontrun{
generar_excel_resultados_15_min(mis_datos, "llr_40", "misma_persona")
}
}
