% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tippett.r
\name{generar_graficos_tippett_con3}
\alias{generar_graficos_tippett_con3}
\title{Generar múltiples gráficas Tippett (versión 3 diferencias)}
\usage{
generar_graficos_tippett_con3(llr_var)
}
\arguments{
\item{llr_var}{Nombre de la columna con valores de LLR}
}
\value{
No retorna valor, muestra las gráficas en una cuadrícula
}
\description{
Genera una cuadrícula de 2x2 con gráficas Tippett para diferencias 0 a 3.
}
\examples{
\dontrun{
generar_graficos_tippett_con3("llr_40")
}
}
