% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecciones.r
\name{graficar_intersecciones}
\alias{graficar_intersecciones}
\title{Graficar intersecciones}
\usage{
graficar_intersecciones(resultados_combinados)
}
\arguments{
\item{resultados_combinados}{Data frame con resultados de intersecciones}
}
\value{
Objeto ggplot
}
\description{
Genera un gráfico de puntos que muestra las intersecciones
para diferentes modelos y niveles de diferencia.
}
\examples{
\dontrun{
graficar_intersecciones(mis_resultados)
}
}
