% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_LR.R
\name{print.calibrated_LR}
\alias{print.calibrated_LR}
\title{Print method for calibrated_LR objects}
\usage{
\method{print}{calibrated_LR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"calibrated_LR"}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\value{
Invisibly returns the object \code{x}.
}
\description{
Print method for calibrated_LR objects
}
