% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_LR.R
\name{summary.calibrated_LR}
\alias{summary.calibrated_LR}
\title{Summary method for calibrated_LR objects}
\usage{
\method{summary}{calibrated_LR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"calibrated_LR"}.}

\item{...}{Additional arguments passed to \code{\link{summary}}.}
}
\value{
Invisibly returns the object \code{object}.
}
\description{
Summary method for calibrated_LR objects
}
