% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecciones.r
\name{transformar_tabla_intersecciones}
\alias{transformar_tabla_intersecciones}
\title{Transformar tabla de intersecciones}
\usage{
transformar_tabla_intersecciones(tabla_intersecciones)
}
\arguments{
\item{tabla_intersecciones}{Data frame con intersecciones}
}
\value{
Data frame transformado
}
\description{
Convierte una tabla de intersecciones a formato ancho
y aplica valor absoluto a los valores.
}
\examples{
\dontrun{
tabla_ancha <- transformar_tabla_intersecciones(mis_intersecciones)
}
}
