% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculos.r
\name{calibrate_LR}
\alias{calibrate_LR}
\title{Calibrar LRs mediante regresión logística}
\usage{
calibrate_LR(data, LR_col = "LR", binary_col = "misma_persona_binaria")
}
\arguments{
\item{data}{Data frame con los datos}

\item{LR_col}{Nombre de la columna con los LRs}

\item{binary_col}{Nombre de la columna binaria (1 = misma persona, 0 = diferente)}
}
\value{
Lista con coeficientes, prior odds, modelo ajustado y LRs calibrados
}
\description{
Aplica calibración a los ratios de verosimilitud (LRs) usando
regresión logística para mejorar su rendimiento.
}
\examples{
\dontrun{
resultado <- calibrate_LR(mis_datos, "LR", "misma_persona_binaria")
lrs_calibrados <- resultado$calibrated_LRs
}
}
